# docs/swagger.yaml
components:
  schemas:
...
    github_com_array_banking-api_internal_errors.ErrorDetail:
      properties:
        code:
          type: string
        details:
          items:
            type: string
          type: array
          uniqueItems: false
        message:
          type: string
        trace_id:
          type: string
      type: object
    github_com_array_banking-api_internal_errors.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorDetail'
      type: object
...
    internal_handlers.SuccessResponse:
      allOf:
      - $ref: '#/components/schemas/message'
      properties:
        data:
          type: object
        message:
          type: string
        meta:
          type: object
      type: object
    message:
      properties:
        message:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
        eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: support@arraybanking.com
    name: API Support
  description: Production-quality banking REST API for developer assessment and interviewing.
    Provides core banking functionality including identity management, account operations,
    customer management, transaction processing, NorthWind Bank integration, and regulator
    compliance notifications.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Array Banking API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /admin/users:
    get:
      description: Admin endpoint to list all users with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Items per page (max 100)
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_handlers.SuccessResponse'
          description: Users retrieved successfully with pagination metadata
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid pagination parameters
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Requires admin role
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - Admin
  /admin/users/{userId}:
    delete:
      description: Admin endpoint to soft delete a user. Cannot delete own account.
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_handlers.SuccessResponse'
          description: User deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid user ID or cannot delete own account
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Requires admin role
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: CUSTOMER_001 - User not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Delete user (admin)
      tags:
      - Admin
    get:
      description: Admin endpoint to retrieve detailed user information
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_handlers.SuccessResponse'
          description: User retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid user ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Requires admin role
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: CUSTOMER_001 - User not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get user by ID (admin)
      tags:
      - Admin
...
  /admin/users/{userId}/unlock:
    post:
      description: Admin endpoint to unlock a locked user account
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/internal_handlers.SuccessResponse'
          description: User unlocked successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid user ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Requires admin role
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: CUSTOMER_001 - User not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Unlock user account (admin)
      tags:
      - Admin
...
servers:
- description: Local development server
  url: http://localhost:8080
- description: Production server
  url: https://api.arraybanking.com